Solution
Your goal is to create a 3x3 table, with proper headings (so it's really 4x3 if you count the heading cells). There are  some starter data in index.html. Make sure you create a 3x3 table with a row of headings at the top, using the correct semantic elements. You are expected to use the following elements: table, thead, tr, th, tbody, and td.